Inroduce ConsensusAlgorithm namespace
